internal/gamepad, internal/glfw: fix unsafe pointer use in objc calls
Passing uintptr(unsafe.Pointer(&x)) to objc.ID.Send hands Objective-C a
plain integer. Go's unsafe rules keep such a value valid only inside a
syscall.Syscall argument list, so the referent is neither kept alive nor
updated when the goroutine stack moves. Escape analysis confirms the
referents stayed on the stack: not one of them moved to the heap.

The window is wide. Every Send boxes its arguments into interfaces and
runs purego's reflection path before reaching objc_msgSend, so a stack
growth or a garbage collection stack shrink can relocate the frame in
between.

The sites whose callee writes through the pointer are the worse case.
The write lands in stack memory that has been freed, and the Go side
afterwards reads a variable the callee never touched, so a failed call
reports no error. That covers the swap interval readback and every
NSError out-parameter.

Pass unsafe.Pointer instead. purego marshals it as a pointer, and the
compiler then keeps the referent alive and updates it across stack
moves. The nib objects out-parameter names a package-level variable that
never moves, and changes only for consistency.

Closes #3701

+++ internal/glfw/cocoa_init_darwin.go
@@ -488,7 +488,7 @@ func platformInit() error {
 						bundle.Send(objc.RegisterName("loadNibNamed:owner:topLevelObjects:"),
 							mainMenuNib.ID,
 							objc.ID(class_NSApplication).Send(sel_sharedApplication),
-							uintptr(unsafe.Pointer(&_glfw.platformWindow.nibObjects)))
+							unsafe.Pointer(&_glfw.platformWindow.nibObjects))
 					} else {
 						createMenuBar()
 					}
